# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is API specification document for Hit it Solution",
        "version": "2.0.0",
        "title": "Hit It Solution"
    },
    "basePath": "/api/hitit",
    "schemes": [
        "http",
        "http"
    ],
    "paths": {
...
        "/register/college": {
            "post": {
                "tags": [
                    "College"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/createCollageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created Successful",
                        "schema": {
                            "$ref": "#/definitions/successResponce"
                        }
                    },
                    "400": {
                        "description": "Invalid Data",
                        "schema": {
                            "$ref": "#/definitions/errorResponce"
                        }
                    }
                }
            }
        },
        "/register/collegehr": {
            "post": {
                "tags": [
                    "College HR"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/createCollageHr"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created Successful",
                        "schema": {
                            "$ref": "#/definitions/successResponce"
                        }
                    },
                    "400": {
                        "description": "Invalid Data",
                        "schema": {
                            "$ref": "#/definitions/errorResponce"
                        }
                    }
                }
            }
        },
        "/register/depthr": {
            "post": {
                "tags": [
                    "College HR"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/createDeptHr"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created Successful",
                        "schema": {
                            "$ref": "#/definitions/successResponce"
                        }
                    },
                    "400": {
                        "description": "Invalid Data",
                        "schema": {
                            "$ref": "#/definitions/errorResponce"
                        }
                    }
                }
            }
        },
        "/register/student": {
            "post": {
                "tags": [
                    "Department HR"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/createStudents"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created Successful",
                        "schema": {
                            "$ref": "#/definitions/successResponce"
                        }
                    },
                    "400": {
                        "description": "Invalid Data",
                        "schema": {
                            "$ref": "#/definitions/errorResponce"
                        }
                    }
                }
            }
        },
        "/register/company": {
            "post": {
                "tags": [
                    "Company"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/companyData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created Successful",
                        "schema": {
                            "$ref": "#/definitions/successResponce"
                        }
                    },
                    "400": {
                        "description": "Invalid Data",
                        "schema": {
                            "$ref": "#/definitions/errorResponce"
                        }
                    }
                }
            }
        },
...
        "/event/attendance": {
            "post": {
                "tags": [
                    "Events"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/eventAttendance"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created Successful",
                        "schema": {
                            "$ref": "#/definitions/successResponce"
                        }
                    },
                    "400": {
                        "description": "Invalid Data",
                        "schema": {
                            "$ref": "#/definitions/errorResponce"
                        }
                    }
                }
            }
        },
...
        "/event/{userId}/{offset}/{totalRecords}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "type": "string",
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "offset",
                        "type": "number",
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "totalRecords",
                        "type": "number",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request Successful",
                        "schema": {
                            "$ref": "#/definitions/eventsResponce"
                        }
                    },
                    "400": {
                        "description": "Invalid User",
                        "schema": {
                            "$ref": "#/definitions/errorResponce"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "eventsResponce": {
            "type": "array",
            "items": {
                "type": "object",
                "properties": {
                    "eventName": {
                        "type": "string"
                    },
                    "eventType": {
                        "type": "string"
                    },
                    "eventDescrib": {
                        "type": "string"
                    },
                    "dateFrom": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "dateTo": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "venue": {
                        "type": "string"
                    }
                }
            }
        },
...
        "eventAttendance": {
            "type": "object",
            "properties": {
                "eventId": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                },
                "attendance": {
                    "type": "string"
                }
            }
        },
...
        "companyData": {
            "type": "object",
            "properties": {
                "companyName": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                },
                "mailId": {
                    "type": "string"
                },
                "companyTech": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "companyBranchs": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "branchName": {
                                "type": "string"
                            },
                            "branchLocationCity": {
                                "type": "string"
                            },
                            "mailId": {
                                "type": "string"
                            },
                            "isheadBranch": {
                                "type": "boolean"
                            },
                            "projects": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "projectName": {
                                            "type": "string"
                                        },
                                        "projectType": {
                                            "type": "string"
                                        },
                                        "projectDescrib": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "createStudents": {
            "type": "object",
            "properties": {
                "userEmail": {
                    "type": "string"
                },
                "collegeId": {
                    "type": "string"
                },
                "deptId": {
                    "type": "string"
                }
            }
        },
        "createDeptHr": {
            "type": "object",
            "properties": {
                "userEmail": {
                    "type": "string"
                },
                "deptId": {
                    "type": "string"
                },
                "collegeId": {
                    "type": "string"
                }
            }
        },
        "createCollageHr": {
            "type": "object",
            "properties": {
                "userEmail": {
                    "type": "string"
                },
                "collegeId": {
                    "type": "string"
                }
            }
        },
        "createCollageRequest": {
            "type": "object",
            "properties": {
                "collegeName": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                },
                "collegeDetail": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "college_email": {
                                "type": "string"
                            },
                            "college_group_email": {
                                "type": "string"
                            },
                            "college_phone": {
                                "type": "number"
                            },
                            "college_address": {
                                "type": "object",
                                "properties": {
                                    "pincode": {
                                        "type": "number"
                                    },
                                    "country": {
                                        "type": "string"
                                    },
                                    "state": {
                                        "type": "string"
                                    },
                                    "district": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "successResponce": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "errorResponce": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}